refactor: refactor `JiraChangelog`
- Move `JiraChangelog` to `/models/jira/common.py`
- Change `JiraChangelog.created` from `str` to `datetime | None`
